Added controller class
